twilio-voice-ai/requirements.txt
flask==3.1.1
twilio==9.6.1
requests==2.32.3
python-dotenv==1.1.0
gunicorn==23.0.0
